# frontend/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Chat Systems API",
    "description": "Every error response carries `{\"error\": \"<message>\"}` with the HTTP status saying what kind of error it is.",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/api"
    }
  ],
  "paths": {
...
    "/channels/{ch_id}/messages": {
      "get": {
        "tags": [
          "messages"
        ],
        "operationId": "messaging_list_messages",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int64"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ],
              "format": "uuid"
            }
          },
          {
            "name": "ch_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataList_MessageWithReactions"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "messages"
        ],
        "operationId": "messaging_send_message",
        "parameters": [
          {
            "name": "ch_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendMessageRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          }
        }
      }
    },
...
    "/messages/{msg_id}": {
      "delete": {
        "tags": [
          "messages"
        ],
        "operationId": "messaging_delete_message",
        "parameters": [
          {
            "name": "msg_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusResponse"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "messages"
        ],
        "operationId": "update_message",
        "parameters": [
          {
            "name": "msg_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMessageRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          }
        }
      }
    },
    "/messages/{msg_id}/history": {
      "get": {
        "tags": [
          "messages"
        ],
        "summary": "The author sees their own prior versions; a workspace admin sees anyone's,\nand that read is audited. Everyone else gets the \"edited\" marker and nothing\nbehind it — showing every earlier draft to every reader is a different\nproduct from the one people think they are using.",
        "operationId": "messaging_message_history",
        "parameters": [
          {
            "name": "msg_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataList_MessageEdit"
                }
              }
            }
          }
        }
      }
    },
...
    "/workspaces/{ws_id}/conversations": {
      "get": {
        "tags": [
          "conversations"
        ],
        "operationId": "list_conversations",
        "parameters": [
          {
            "name": "ws_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataList_ConversationSummary"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "conversations"
        ],
        "operationId": "create_conversation",
        "parameters": [
          {
            "name": "ws_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateConversationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Conversation"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BotIdentity": {
        "type": "object",
        "required": [
          "hook_id",
          "name"
        ],
        "properties": {
          "hook_id": {
            "type": "string",
            "format": "uuid"
          },
          "icon_url": {
            "type": [
              "string",
              "null"
            ]
          },
          "name": {
            "type": "string"
          }
        }
      },
...
      "Conversation": {
        "type": "object",
        "required": [
          "id",
          "workspace_id",
          "kind",
          "last_message_at",
          "created_at"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "created_by": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "kind": {
            "$ref": "#/components/schemas/ConversationKind"
          },
          "last_message_at": {
            "type": "string",
            "format": "date-time"
          },
          "workspace_id": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "ConversationKind": {
        "type": "string",
        "description": "A direct message is the two-person case of a group; both are channels of a\ntype nobody can browse or join, and this is the name the client knows them by.",
        "enum": [
          "direct",
          "group"
        ]
      },
...
      "CreateConversationRequest": {
        "type": "object",
        "required": [
          "participant_ids"
        ],
        "properties": {
          "participant_ids": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          }
        }
      },
...
      "DataList_ConversationSummary": {
        "type": "object",
        "required": [
          "data"
        ],
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "required": [
                "id",
                "workspace_id",
                "kind",
                "last_message_at",
                "participant_ids"
              ],
              "properties": {
                "id": {
                  "type": "string",
                  "format": "uuid"
                },
                "kind": {
                  "$ref": "#/components/schemas/ConversationKind"
                },
                "last_message_at": {
                  "type": "string",
                  "format": "date-time"
                },
                "last_read_at": {
                  "type": [
                    "string",
                    "null"
                  ],
                  "format": "date-time"
                },
                "participant_ids": {
                  "type": "array",
                  "items": {
                    "type": "string",
                    "format": "uuid"
                  }
                },
                "workspace_id": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          }
        }
      },
...
      "DataList_MessageEdit": {
        "type": "object",
        "required": [
          "data"
        ],
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "required": [
                "id",
                "message_id",
                "previous_content",
                "edited_by",
                "edited_at"
              ],
              "properties": {
                "edited_at": {
                  "type": "string",
                  "format": "date-time"
                },
                "edited_by": {
                  "type": "string",
                  "format": "uuid"
                },
                "id": {
                  "type": "string",
                  "format": "uuid"
                },
                "message_id": {
                  "type": "string",
                  "format": "uuid"
                },
                "previous_content": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "DataList_MessageWithReactions": {
        "type": "object",
        "required": [
          "data"
        ],
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/Message"
                },
                {
                  "type": "object",
                  "required": [
                    "reactions"
                  ],
                  "properties": {
                    "reactions": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Reaction"
                      }
                    }
                  }
                }
              ]
            }
          }
        }
      },
...
      "Message": {
        "type": "object",
        "required": [
          "id",
          "channel_id",
          "user_id",
          "content",
          "metadata",
          "reply_count",
          "is_pinned",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "channel_id": {
            "type": "string",
            "format": "uuid"
          },
          "client_message_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "content": {
            "type": "string"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "deleted_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "is_pinned": {
            "type": "boolean"
          },
          "metadata": {
            "$ref": "#/components/schemas/MessageMetadata"
          },
          "reply_count": {
            "type": "integer",
            "format": "int32"
          },
          "thread_parent_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          },
          "user_id": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
...
      "MessageMetadata": {
        "type": "object",
        "properties": {
          "bot": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/BotIdentity"
              }
            ]
          },
          "huddle_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "initiator_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "kind": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "Reaction": {
        "type": "object",
        "required": [
          "id",
          "message_id",
          "user_id",
          "emoji",
          "created_at"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "emoji": {
            "type": "string"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "message_id": {
            "type": "string",
            "format": "uuid"
          },
          "user_id": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
...
      "SendMessageRequest": {
        "type": "object",
        "required": [
          "content"
        ],
        "properties": {
          "client_message_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid",
            "description": "The sender's own id for this send, used only to make a retry idempotent.\nIt is not the message id — the server owns that."
          },
          "content": {
            "type": "string"
          },
          "thread_parent_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          }
        }
      },
...
      "StatusResponse": {
        "type": "object",
        "required": [
          "status"
        ],
        "properties": {
          "status": {
            "type": "string"
          }
        }
      },
...
      "UpdateMessageRequest": {
        "type": "object",
        "required": [
          "content"
        ],
        "properties": {
          "content": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Same-origin clients are authenticated by the httpOnly cookie instead."
      }
    }
  },
  "security": [
    {
      "bearer": []
    }
  ],
...
}